# swagger-spec-account-op.json
{"openapi":"3.0.0","info":{"title":"Open Platform Service Outlook Barbershop","description":"This API Docs for describe all Open Platform RESTful endpoint","version":"V 1.0","contact":{"name":"Alphacreativee","url":"http://alphacreativee.com","email":"support@alphacreativee.co.id"}},"tags":[],"servers":[{"url":"http://localhost:3000","description":"Development API Gateway"}],"components":{"schemas":{"V1CreateBarbermansDto":{"type":"object","properties":{"name":{"type":"string","example":"Jackson","description":"Barber man Name"},"status":{"type":"number","example":1,"description":"Status of Barber man","enum":[0,1]},"price":{"type":"number","example":35000,"description":"Barber man Net Price"},"price_discount":{"type":"number","example":30000,"description":"Barber man Discount Price"},"location_id":{"type":"number","example":1,"description":"Barber man Location Ids"},"service":{"type":"number","example":1,"description":"Barber man Status Service"},"email":{"type":"string","example":"jackson@gmail.com","description":"Barber man email"},"photo":{"type":"object","example":"data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEA","description":"Base64 image"}},"required":["name","status","price","price_discount","location_id","service","email","photo"]},"Error400":{"type":"object","properties":{"status":{"type":"number","description":"Status Http Code","example":400},"message":{"type":"string","description":"Error Message","example":"<Property> should not be empty"},"type":{"type":"string","description":"Error Type Name of Http Status","example":"Bad Request"},"code":{"type":"number","description":"Custom Code for Error","example":null}},"required":["status","message","type","code"]},"Error401":{"type":"object","properties":{"status":{"type":"number","description":"Status Http Code","example":401},"message":{"type":"string","description":"Error Message","example":"Unauthorized"},"type":{"type":"string","description":"Error Type Name of Http Status","example":"Unauthorized"},"code":{"type":"number","description":"Custom Code for Error","example":null}},"required":["status","message","type","code"]},"Error403":{"type":"object","properties":{"status":{"type":"number","description":"Status Http Code","example":403},"message":{"type":"string","description":"Error Message","example":"This <Property> Phone already exist"},"type":{"type":"string","description":"Error Type Name of Http Status","example":"Forbidden"},"code":{"type":"number","description":"Custom Code for Error","example":null}},"required":["status","message","type","code"]},"Error404":{"type":"object","properties":{"status":{"type":"number","description":"Status Http Code","example":404},"message":{"type":"string","description":"Error Message","example":"<Property> is Not Found"},"type":{"type":"string","description":"Error Type Name of Http Status","example":"Not Found"},"code":{"type":"number","description":"Custom Code for Error","example":null}},"required":["status","message","type","code"]},"Error500":{"type":"object","properties":{"status":{"type":"number","description":"Status Http Code","example":500},"message":{"type":"string","description":"Error Message","example":"Something Wrong, Please Try Again"},"type":{"type":"string","description":"Error Type Name of Http Status","example":"Internal Server Error"},"code":{"type":"number","description":"Custom Code for Error","example":null}},"required":["status","message","type","code"]},"BarbermansResponseData":{"type":"object","properties":{"name":{"type":"string","example":"Jackson"},"status":{"type":"number","example":1},"price":{"type":"number","example":35000},"price_discount":{"type":"number","example":30000},"location_id":{"type":"number","example":1},"service":{"type":"number","example":1},"email":{"type":"string","example":"jackson@gmail.com"},"photo":{"type":"string","example":"data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEA"}},"required":["name","status","price","price_discount","location_id","service","email","photo"]},"Meta":{"type":"object","properties":{"skip":{"type":"number","description":"Amount Skip Data from list","example":0},"limit":{"type":"number","description":"Amount Total Data from list per page","example":10},"total":{"type":"number","description":"Amount total data","example":1}},"required":["skip","limit","total"]},"SuccessBarbermanFindDoc":{"type":"object","properties":{"data":{"description":"Wrapper Message","type":"array","items":{"$ref":"#/components/schemas/BarbermansResponseData"}},"meta":{"description":"Meta Data","allOf":[{"$ref":"#/components/schemas/Meta"}]}},"required":["data","meta"]}}},"paths":{"/api/v1/barbermans":{"post":{"operationId":"V1BarbermanController_create","summary":"(Create Barberman)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/V1CreateBarbermansDto"}}}},"responses":{"200":{"description":"Success"},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error400"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error401"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error403"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error404"}}}},"500":{"description":"General Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error500"}}}}},"tags":["Barberman Endpoint List"]},"get":{"operationId":"V1BarbermanController_find","summary":"(List Barbermans)","parameters":[{"name":"skip","required":true,"in":"query","example":"0","description":"Offset Data from List Pagination","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","example":"10","description":"Limit Data from List Pagination","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","example":"asc","description":"Type of sort data","schema":{"enum":["asc","desc","ASC","DESC"],"type":"string"}},{"name":"sortby","required":false,"in":"query","example":"id","description":"Key of sort data","schema":{"enum":["id","name","created_at"],"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessBarbermanFindDoc"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error400"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error401"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error403"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error404"}}}},"500":{"description":"General Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error500"}}}}},"tags":["Barberman Endpoint List"]}}}}